Also rename folders
