+++ packages/data-with-position/src/__fixtures__/from-yaml/basic.yaml
