+++ packages/data-with-position/src/__fixtures__/from-yaml/leading newline.config.ts
@@ -1,4 +1,4 @@
-import { FromYamlTestCaseConfig } from "../types";
+import type { FromYamlTestCaseConfig } from "../types";
 
 export const config: FromYamlTestCaseConfig = {
   nodesToCheck: [
