+++ packages/literate-elm/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./codeInsights";
+export * from "./code-insights";
 export * from "./ensure-environment";
 export * from "./run-program";
 export * from "./types";
