+++ packages/litvis-integration-mume/src/index.ts
@@ -1,6 +1,6 @@
 export * from "./init-litvis-enhancer-cache";
-export { useMarkdownItLitvisFeatures } from "./markdownItFeatures";
-export * from "./renderEnhancers";
-export * from "./renderPostEnhancers";
+export { useMarkdownItLitvisFeatures } from "./markdown-it-features";
+export * from "./render-enhancers";
+export * from "./render-post-enhancers";
 export * from "./types";
 export { loadAndProcessLitvisNarrative } from "litvis";
