+++ packages/litvis/src/attribute-derivatives/extract-attribute-derivatives.ts
