+++ packages/litvis/src/attribute-derivatives/resolve-expressions.ts
