+++ packages/litvis/src/document/process-frontmatter/process-frontmatter.ts
