+++ packages/litvis/src/index.ts
@@ -1,7 +1,7 @@
 export {
   extractAttributeDerivatives,
   resolveExpressions,
-} from "./attributeDerivatives";
+} from "./attribute-derivatives";
 export * from "./init-cache";
 export * from "./narrative/list-narrative-files";
 export * from "./narrative/load-and-process-litvis-narrative";
